REFACTOR: Window component moved to Window folder
