+++ components/Window/window.tsx
