+++ components/customWindows/myCertifications.tsx
@@ -1,6 +1,6 @@
 import { certificatesData } from "@/data/certificatesData";
 import Link from "next/link";
-import Window from "../window";
+import Window from "../Window/window";
 
 export default function MyCertificationsWindow(): JSX.Element {
